A6: split AudienceRepository into Reader + Writer behind a façade (#563 backlog)
AudienceReader holds the 17 read queries (hierarchy traversal, membership,
search, counts); AudienceWriter holds the 9 writes (create/update/delete,
member mutations, cascade_self_join). Both use StaticRepositoryTrait with the
same cache group so writes still invalidate read caches; the writer cross-calls
the reader for the reads its write paths need (get_children, is_member).
AudienceRepository stays the public façade delegating every method, with the
@phpstan-type and a tech-debt note (#563 B3).

Behavior-preserving — AudienceRepositoryTest (79) + Audience suite (652) green;
PHPStan L8 + WPCS clean.
